Added auto-registration in monitoring service.

+++ src/Lykke.Service.HFT/Models/ClientBalanceResponseModel.cs
@@ -3,9 +3,9 @@
     public class ClientBalanceResponseModel
     {
         public string AssetId { get; set; }
-        public double Balance { get; set; }
-        public double Reserved { get; set; }
-        
+        public decimal Balance { get; set; }
+        public decimal Reserved { get; set; }
+
         public static ClientBalanceResponseModel Create(Balances.AutorestClient.Models.ClientBalanceResponseModel src)
         {
             return new ClientBalanceResponseModel
